opendbc.can: split out DBC (#2543)
* opendbc.can: split out DBC

* comment

* fix that

* one more
